apps/backend/requirements.txt
litellm==1.66.1
fastapi==0.108.0
uvicorn==0.27.0
pydantic==2.5.3
scikit-learn==1.4.0
pytest==7.4.3
pandas==2.1.4
numpy==1.26.3
rank_bm25==0.2.2
python-multipart==0.0.9
python-dotenv==1.0.0
requests==2.31.0 